+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsejsonschema.py
@@ -184,9 +184,9 @@ class EvaluatorResponseJSONSchema(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-29T09:12:37.464Z"
+    created: Optional[str] = "2026-06-29T13:49:50.653Z"
 
-    updated: Optional[str] = "2026-06-29T09:12:37.464Z"
+    updated: Optional[str] = "2026-06-29T13:49:50.653Z"
 
     updated_by_id: OptionalNullable[str] = UNSET