+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsetypescript.py
@@ -186,9 +186,9 @@ class EvaluatorResponseTypescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-29T09:12:37.464Z"
+    created: Optional[str] = "2026-06-29T13:49:50.653Z"
 
-    updated: Optional[str] = "2026-06-29T09:12:37.464Z"
+    updated: Optional[str] = "2026-06-29T13:49:50.653Z"
 
     updated_by_id: OptionalNullable[str] = UNSET