+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KW9AEEN3C8SG0JKG3XR51CJP"
+        "01KW9TA5DNS85MTKZPQ0VE2H3Q"
     )
     r"""The unique identifier of the data source"""