+++ packages/orq-rc/RELEASES.md
@@ -9049,3 +9049,13 @@ Based on:
 - [python v4.11.0-rc.54] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.54 - packages/orq-rc
+
+## 2026-06-29 13:52:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.55] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.55] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.55 - packages/orq-rc